make carbon schema compatible with hive

+++ integration/hive/src/main/java/org/apache/carbondata/hive/CarbonHiveRecordReader.java
@@ -224,11 +224,11 @@ public class CarbonHiveRecordReader extends CarbonRecordReader<ArrayWritable>
       case DOUBLE:
         return new DoubleWritable((double) obj);
       case INT:
-        return new IntWritable(((Long) obj).intValue());
+        return new IntWritable(((int) obj));
       case LONG:
         return new LongWritable((long) obj);
       case SHORT:
-        return new ShortWritable(((Long) obj).shortValue());
+        return new ShortWritable(((short) obj));
       case DATE:
         return new DateWritable(new Date(Long.parseLong(String.valueOf(obj.toString()))));
       case TIMESTAMP:
